requirements.txt
amqp==5.3.1
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-filter==25.2
django-timezone-field==7.2.1
djangorestframework==3.17.1
idna==3.13
kombu==5.6.2
lxml==6.1.0
packaging==26.2
prompt_toolkit==3.0.52
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.4.0
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
